ResultT transformer + tests
